--- apps/dash-brain-viewer/.gitignore
@@ -1,4 +1,0 @@
-venv
-*.pyc
-.DS_Store
-.env
