+++ apps/dash-brain-viewer/assets/images/brain.png
