+++ apps/dash-brain-viewer/constants.py
@@ -1,0 +1,34 @@
+from mni import create_mesh_data, default_colorscale
+import os
+
+
+GITHUB_LINK = os.environ.get(
+    "GITHUB_LINK",
+    "https://github.com/plotly/dash-sample-apps/tree/master/apps/dash-brain-viewer",
+)
+
+default_colorscale_index = [ea[1] for ea in default_colorscale]
+
+axis_template = {
+    "showbackground": True,
+    "backgroundcolor": "#141414",
+    "gridcolor": "rgb(255, 255, 255)",
+    "zerolinecolor": "rgb(255, 255, 255)",
+}
+
+plot_layout = {
+    "title": "",
+    "margin": {"t": 0, "b": 0, "l": 0, "r": 0},
+    "font": {"size": 12, "color": "white"},
+    "showlegend": False,
+    "plot_bgcolor": "#141414",
+    "paper_bgcolor": "#141414",
+    "scene": {
+        "xaxis": axis_template,
+        "yaxis": axis_template,
+        "zaxis": axis_template,
+        "aspectratio": {"x": 1, "y": 1.2, "z": 1},
+        "camera": {"eye": {"x": 1.25, "y": 1.25, "z": 1.25}},
+        "annotations": [],
+    },
+}
