+++ apps/dash-brain-viewer/requirements.txt
@@ -1,4 +1,4 @@
-dash==1.0.2
-gunicorn==19.9.0
-pandas==0.24.2
+dash==2.4.1
+pandas==1.4.2
+gunicorn==20.1.0
 dash-colorscales==0.0.4
